+++ TMP006/TMP006.ino
@@ -1,67 +1,72 @@
-/*************************************************** 
-  This is an example for the TMP006 Contact-less Infrared Thermopile Sensor 
+// Headers for each sensor type
+#include "src/TMP006.h" 
+//...
 
-  Designed specifically to work with the Adafruit TMP006 Breakout 
-  ----> https://www.adafruit.com/products/1296
+#include "src/Sensor.h"
+#define NUMSENSORS 1 //Or however many
+#define BAUDRATE 115200
+#define THISARDUINO ARDUINO_ONE
 
-  These displays use I2C to communicate, 2 pins are required to  
-  interface
-  Adafruit invests time and resources providing this open source code, 
-  please support Adafruit and open-source hardware by purchasing 
-  products from Adafruit!
+// Objects for each sensor
+TMP006 tmp006(TMP006_I2CADDR, THISARDUINO); 
+//...
 
-  Written by Limor Fried/Ladyada for Adafruit Industries.  
-  BSD license, all text above must be included in any redistribution
- ****************************************************/
+Sensor* sensors[NUMSENSORS] = {
+    // Entry for each sensor object
+    &tmp006, 
+    
+    //...
+};
 
-#include <Wire.h>
-#include <Adafruit_Sensor.h>
-#include "Adafruit_TMP006.h"
+// !#!#!#!--- EVERYTHING AFTER HERE DOES NOT NEED TO BE CHANGED FOR SENSOR IMPLEMENTATION ---!#!#!#!
 
-// Connect VCC to +3V (its a quieter supply than the 5V supply on an Arduino
-// Gnd -> Gnd
-// SCL connects to the I2C clock pin. On newer boards this is labeled with SCL
-// otherwise, on the Uno, this is A5 on the Mega it is 21 and on the Leonardo/Micro digital 3
-// SDA connects to the I2C data pin. On newer boards this is labeled with SDA
-// otherwise, on the Uno, this is A4 on the Mega it is 20 and on the Leonardo/Micro digital 2
+void setup(){
+    Serial.begin(BAUDRATE);
 
-Adafruit_TMP006 tmp006;
-//Adafruit_TMP006 tmp006(0x41);  // start with a diferent i2c address!
+    bool success = true;
+    for(int i = 0; i < NUMSENSORS; i++){
+        SensorState* state = sensors[i]->begin();
 
-void setup() { 
-  Serial.begin(9600);
-  Serial.println("Adafruit TMP006 example");
-
-  // you can also use tmp006.begin(TMP006_CFG_1SAMPLE) or 2SAMPLE/4SAMPLE/8SAMPLE to have
-  // lower precision, higher rate sampling. default is TMP006_CFG_16SAMPLE which takes
-  // 4 seconds per reading (16 samples)
-  if (! tmp006.begin(TMP006_CFG_4SAMPLE)) { // Always takes 4 samples per second
-    Serial.println("No sensor found");
-    while (1);
-  }
-
-  Serial.println("Send s to enter sleep mode, or w to wake up.  Measurements are not updated while asleep!");
+        // Print/send sensor post-setup state data here. For example:
+        bool _success = (state->error == ERR_NONE);
+        if(_success){
+            Serial.print("Sensor ");
+            Serial.print(sensors[i]->sensor);
+            Serial.println(" initialized.");
+        } else {
+            Serial.print("Sensor ");
+            Serial.print(sensors[i]->sensor);
+            Serial.println(" failed to initialize!");
+        }
+        success &= _success;
+    }
+    if(!success){
+        Serial.println("POST failed on one or more sensors, freezing...");
+        while(1){delay(1000);}
+    }
 }
 
-void loop() {
-  // Check for sleep/wake command.
-  while (Serial.available() > 0) {
-    char c = Serial.read();
-    if (c == 'w') {
-      Serial.println("Waking up!");
-      tmp006.wake();
+void loop(){
+    for(int i = 0; i < NUMSENSORS; i++){
+        SensorState* state = sensors[i]->update();
+        // Print/send sensor post-setup state data here. For example:
+        bool _success = (state->error == ERR_NONE);
+        bool _new = (state->debug == DS_NEWREAD);
+        if(_success && _new){
+            Serial.print("Sensor ");
+            Serial.print(sensors[i]->sensor);
+            Serial.print(" read success: ");
+            for(int x = 0; x < state->numdata; x++){
+                Serial.print(state->data[x].data);
+                Serial.print(' ');
+                Serial.print(state->data[x].units);
+                if(x < state->numdata-1){Serial.print(", ");}
+            }
+        } else {
+            Serial.print("Sensor ");
+            Serial.print(sensors[i]->sensor);
+            Serial.println(" failed to update!");
+            // TODO: Recover failed sensor?
+        }
     }
-    else if (c == 's') {
-      Serial.println("Going to sleep!");
-      tmp006.sleep();
-    }
-  }
-
-  // Grab temperature measurements and print them.
-  float objt = tmp006.readObjTempC();
-  Serial.print("Object Temperature: "); Serial.print(objt); Serial.println("*C");
-  float diet = tmp006.readDieTempC();
-  Serial.print("Die Temperature: "); Serial.print(diet); Serial.println("*C");
-   
-  delay(1000);
 }
