+++ TMP006/src/Sensor.h
