+++ TMP006/src/TMP006.h
@@ -1,16 +1,16 @@
 #ifndef TMP006_H
 #define TMP006_H
 
-#include "../Sensor.h"
+#include "Sensor.h"
 #include <Wire.h>
 #include <Adafruit_Sensor.h>
-#include "libraries/Adafruit_TMP006.h"
+#include "Adafruit_TMP006.h"
 
 class TMP006 : public Sensor {
     public:
-        TMP006(uint8_t addr = TMP006_I2CADDR, arduino_t arduino);
+        TMP006(uint8_t addr, arduino_t arduino);
     private:
-        Adafruit_TMP006 tmp006;
+        Adafruit_TMP006* tmp006;
         errorlevel_t read(t_datum* data, uint8_t numdata) override;
         errorlevel_t init() override;
 };
